requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
chardet==3.0.4
discord.py==1.4.1
dynaconf==3.1.0
idna==2.10
loguru==0.5.1
multidict==4.7.6
selenium==3.141.0
urllib3==1.25.10
yarl==1.5.1
